projects/job-aggregator/pyproject.toml
# =============================================================================
# pyproject.toml — Job Aggregator
# =============================================================================
# Single source of truth for project metadata, dependencies, and tool config.
# Install for development:  pip install -e ".[dev]"
# Run the server:           python -m job_aggregator
# Run tests:                pytest
# =============================================================================

[build-system]
requires = ["setuptools>=68.0", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "job-aggregator"
version = "0.1.0"
description = "Scam-free job aggregator — verified sources, W2 filtering, trust scoring"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

# ---------------------------------------------------------------------------
# Runtime dependencies
# ---------------------------------------------------------------------------
dependencies = [
    "fastapi>=0.115",          # Async web framework
    "uvicorn[standard]>=0.34", # ASGI server with auto-reload
    "sqlalchemy[asyncio]>=2.0", # Async ORM for PostgreSQL
    "asyncpg>=0.30",           # Fast async PostgreSQL driver
    "alembic>=1.14",           # Database migration tool
    "redis>=5.0",              # Redis client (with hiredis for speed)
    "hiredis>=3.0",            # C parser for Redis protocol
    "pydantic-settings>=2.0",  # Config from environment variables
    "httpx>=0.28",             # Async HTTP client for external APIs
    "feedparser>=6.0",         # RSS/Atom feed parser
    "apscheduler>=3.10",       # Background job scheduler
]

# ---------------------------------------------------------------------------
# Optional dependency groups
# ---------------------------------------------------------------------------
[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "ruff>=0.4",
    "mypy>=1.8",
    "httpx>=0.28",             # Also used for FastAPI TestClient
]

# ---------------------------------------------------------------------------
# CLI entry point
# ---------------------------------------------------------------------------
[project.scripts]
job-aggregator = "job_aggregator.__main__:main"

# ---------------------------------------------------------------------------
# Tool config
# ---------------------------------------------------------------------------

[tool.ruff]
src = ["src"]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "ASYNC", # flake8-async
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
check_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.setuptools.packages.find]
where = ["src"]
